Add files via upload
This is the project done by Sql structured language. This project is about the Hospital Management system in which it takes how many doctors are there ,hoe many patients are there, when are the appointments, and the prescription give by doctor to patients. This is the theme of the project.